Move decorator to iris helpers
