+++ esmvalcore/preprocessor/_cycles.py
@@ -5,7 +5,7 @@ import logging
 import iris
 import iris.coord_categorisation
 
-from esmvalcore.preprocessor._shared import ignore_iris_vague_metadata_warnings
+from esmvalcore.iris_helpers import ignore_iris_vague_metadata_warnings
 
 logger = logging.getLogger(__name__)
