+++ esmvalcore/preprocessor/_derive/toz.py
@@ -7,7 +7,7 @@ import iris
 from scipy import constants
 
 from esmvalcore.cmor.table import CMOR_TABLES
-from esmvalcore.preprocessor._shared import ignore_iris_vague_metadata_warnings
+from esmvalcore.iris_helpers import ignore_iris_vague_metadata_warnings
 
 from .._regrid import extract_levels, regrid
 from ._baseclass import DerivedVariableBase
