+++ esmvalcore/preprocessor/_derive/troz.py
@@ -3,7 +3,7 @@
 import dask.array as da
 import iris
 
-from esmvalcore.preprocessor._shared import ignore_iris_vague_metadata_warnings
+from esmvalcore.iris_helpers import ignore_iris_vague_metadata_warnings
 
 from ._baseclass import DerivedVariableBase
 from .soz import STRATOSPHERIC_O3_THRESHOLD
