+++ esmvalcore/preprocessor/_derive/uajet.py
@@ -4,7 +4,7 @@ import cf_units
 import iris
 import numpy as np
 
-from esmvalcore.preprocessor._shared import ignore_iris_vague_metadata_warnings
+from esmvalcore.iris_helpers import ignore_iris_vague_metadata_warnings
 
 from ._baseclass import DerivedVariableBase
