+++ esmvalcore/preprocessor/_multimodel.py
@@ -25,11 +25,13 @@ from iris.cube import Cube, CubeList
 from iris.exceptions import MergeError
 from iris.util import equalise_attributes, new_axis
 
-from esmvalcore.iris_helpers import date2num
+from esmvalcore.iris_helpers import (
+    date2num,
+    ignore_iris_vague_metadata_warnings,
+)
 from esmvalcore.preprocessor._shared import (
     _group_products,
     get_iris_aggregator,
-    ignore_iris_vague_metadata_warnings,
 )
 from esmvalcore.preprocessor._supplementary_vars import (
     remove_supplementary_variables,
