+++ tests/unit/preprocessor/test_shared.py
@@ -23,7 +23,6 @@ from esmvalcore.preprocessor._shared import (
     get_array_module,
     get_coord_weights,
     get_iris_aggregator,
-    ignore_iris_vague_metadata_warnings,
     preserve_float_dtype,
     try_adding_calculated_cell_area,
 )
@@ -681,18 +680,3 @@ def test_get_coord_weights_triangular_bound_fail():
     )
     with pytest.raises(ValueError, match=msg):
         get_coord_weights(cube, "latitude")
-
-
-def test_ignore_iris_vague_metadata_warnings():
-    """Test ``ignore_iris_vague_metadata_warnings``."""
-
-    @ignore_iris_vague_metadata_warnings
-    def func():
-        # Collapse non-contiguous coord to check if warning has been raised
-        x_coord = DimCoord([0, 3], bounds=[[-1, 1], [2, 4]], var_name="x")
-        cube = Cube([1, 1], dim_coords_and_dims=[(x_coord, 0)])
-        return cube.collapsed("x", iris.analysis.MEAN)
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("error")
-        func()
