+++ tests/unit/test_iris_helpers.py
@@ -1,12 +1,14 @@
 """Tests for :mod:`esmvalcore.iris_helpers`."""
 
 import datetime
+import warnings
 from copy import deepcopy
 from itertools import permutations
 from pprint import pformat
 from unittest import mock
 
 import dask.array as da
+import iris.analysis
 import numpy as np
 import pytest
 from cf_units import Unit
@@ -26,6 +28,7 @@ from esmvalcore.iris_helpers import (
     has_irregular_grid,
     has_regular_grid,
     has_unstructured_grid,
+    ignore_iris_vague_metadata_warnings,
     merge_cube_attributes,
     rechunk_cube,
     safe_convert_units,
@@ -628,3 +631,18 @@ def test_safe_convert_units(
     assert new_cube is cube
     assert new_cube.standard_name == new_standard_name
     assert new_cube.units == new_units
+
+
+def test_ignore_iris_vague_metadata_warnings():
+    """Test ``ignore_iris_vague_metadata_warnings``."""
+
+    @ignore_iris_vague_metadata_warnings
+    def func():
+        # Collapse non-contiguous coord to check if warning has been raised
+        x_coord = DimCoord([0, 3], bounds=[[-1, 1], [2, 4]], var_name="x")
+        cube = Cube([1, 1], dim_coords_and_dims=[(x_coord, 0)])
+        return cube.collapsed("x", iris.analysis.MEAN)
+
+    with warnings.catch_warnings():
+        warnings.simplefilter("error")
+        func()
